Remove references to test execution review decision or comment
